Refactoring
- not yet happy with how it looks like, but at least allows for testing;
- refactored the daemons and their init process;
- refactored the daemon inputs;

+++ mpnetv2/ai/reader/text_item.py
@@ -63,6 +63,6 @@ class TextItem:
             self.key,
             mapping={
                 **self.asdict(),
-                "vector": vector.tobytes(),
+                "vector": vector.astype(np.float32).tobytes(),
             },
         )
